Rearrange compose networks
The internal network now excludes nginx since it doesn't need to see
most of the other services.

The "external" network is renamed "proxy", and it's only for the rails
app and the nginx server.

nginx can only see Rails now, but Rails is on both networks, so it can
see everything.